[TNF] Rendering pcsd-token uuid to its own machine-config during bootstrapping & plumbing for node info.
This patch adds a template for a pacemaker token that gets rendered during the bootstrap step.
It also adds plumbing to start adding node information from a node lister to the template renderer
for the rendering masters machine config. It should be possible to use these values to update
the pcs auth script to lookup the node names and IPs once MCO is running in controller mode
instead of bootstrap mode.